titanic/requirements.txt
django==3.0.3
djangorestframework==3.13.1
numpy==1.20.3
scikit-learn==0.24.2
pandas==1.2.4